Plot a score overlaid on a supplied t-SNE or PCA generated using a cancer single-cell population expression matrix.

This function is a generic function for plotting a t-SNE or PCA previously generated using cancer single-cell population expression matrix. It can be used to label or colour individual or small groups of cells to assist identification of subpopulations of interest, e.g. groups of cancer stem cells.

plot_scores(
  myData,
  xAxisName,
  yAxisName,
  runName,
  scoresName,
  datasetName,
  outDir,
  labelPositions = NULL,
  colourPositions = NULL,
  highestOnTop = NULL
)

Arguments

myData
Data frame describing t-SNE or PCA and score/s including the score to print.

xAxisName
Name of column in myData that describes x-axis info for plot e.g. "Y1" in a t-SNE.

yAxisName
Name of column in myData that describes y-axis info for plot e.g. "Y2" in a t-SNE.

runName
Run name, e.g. 'tSNE_070'; will be used in the title of the plot and the output tif name.

scoresName
Name of column in myData that contains the score info for plotting e.g. 'ranger_hipsci_raw_IPS'.

datasetName
Name of dataset, e.g. 'mela_4558_TPM'.

outDir
Output directory to use.

labelPositions
Optional: Row number/s of myData i.e. cell positions, to print cell labels for.

colourPositions
Optional: Row number/s of myData i.e. cell positions, to highlight cells (in black).

highestOnTop
Optional: Set to TRUE to print highest scoring cells on top of lower scoring cells.

Value

No return value.