Print expression bar graph for genes/transcripts most highly expressed in selected putative CSCs.

This function prints a bar graph to display expression in the genes most highly expressed in a selected set of putative CSCs. The order is by mean transcript expression in the pCSC cells i.e. highest first. If the no_ribo parameter is used, the same bar graph is printed with all transcripts with "ribosomal" in their full name removed from the results.

printBarchartSingleTranscriptNoStacking(
  tExprMatr,
  pCSC_group_positions,
  pCSC_group_name,
  outDir,
  tissue_type,
  family_type_name,
  patient_source_name,
  no_ribo = TRUE,
  number_of_transcripts_to_show = NULL,
  HUGO_abbreviations_and_fullnames = NULL,
  HUGOFullNameFilesDir = NULL
)

Arguments

tExprMatr
Expression matrix describing a cancer single-cell population; column names should be gene names, row names should be cancer cell names.

pCSC_group_positions
Row number of CSCs in the CSC group in the expression matrix tExprMatr.

pCSC_group_name
Name of CSC group to print preferential expression plots for e.g. 'pCSC_group_1_SC4_Talazoparib.LB17011', but can be anything.

outDir
Output directory for plots.

tissue_type
Name of tissue type e.g. 'CDX'; can be anything.

family_type_name
Name of the gene group type; usually 'HUGO'.

patient_source_name
Optional: Name of the patient sample that CSC cells are from; just name for in the plot title, can be anything.

no_ribo
Optional: Set to 1 to remove ribosomal transcripts from some of the bar charts

number_of_transcripts_to_show
Optional: Number of transcripts to include.

Value

Nothing returned.