Print expression stacked bar graph for HUGO gene groups most preferentially expressed in selected putative CSCs.

This function prints a stacked bar graph to display expression in HUGO gene groups most preferentially expressed in a selected set of putative CSCs. This may use expression in the form of read counts, TPM, TPM count fractions, Cells Per Million, etc. Cells Per Million or TPM count fractions are recommended for tumour cell populations containing a mix of cell types. Read counts are recommended for homogenous cell populations e.g. PDXs, CDXs. The height of each bar is the average expression value for the members of a HUGO gene group in the cells of the selected pCSC group. Each bar is subdivided by the proportion of the gene group average expression contributed by each gene; subdivisions are labelled with gene names, and where the subdivision is too small to fit the gene name in the graphic they are labelled nearby. Bars are ordered by the HUGO gene group average expression value for the pCSC group minus the gene group average expression in all tumour cells from all patients; the difference between the pCSC group average expression and all tumour cells average expression is shown by a pink dashed line. Other horizontal lines: the pCSC group’s source patient average expression value for all tumour cells = dashed green line; average expression value for tumour cells from all patients = red dashed line; all cells from all patients (e.g. including immune cells) average expression value = blue dashed line; and the average expression value for all transcripts in the pCSC group = black dotted line. N.B. not all genes officially defined in a HUGO gene group are present in the scRNA-Seq dataset, depending on the experimental platform used.

stackedFamilyExpressionBarchartSingleBars(
  tExprMatr,
  pCSC_group_positions,
  tumour_only_positions,
  this_patient_tumour_only_positions,
  HUGO_groups_sets_of_columnNumbers,
  pCSC_group_name,
  outDir,
  tissue_type,
  family_type_name,
  patient_source_name,
  frag_count_method,
  separate_row_for_high_expression_transcripts = NULL,
  manually_set_max_number_of_graph_rows = NULL
)

Arguments

tExprMatr
Expression matrix describing a cancer single-cell population; column names should be gene names, row names should be cancer cell names.

pCSC_group_positions
Row number of CSCs in the CSC group in the expression matrix tExprMatr.

tumour_only_positions
Positions (row numbers) in the expression matrix tExprMatr thought to be tumour cells e.g. when the population contains cells annotated or predicted as other cell types a subset may be defined, or if all cells are thought to be malignant cancer cells this may be all cells.

this_patient_tumour_only_positions
Optional: Expression matrix tExprMatr positions (row numbers) for tumour cells from a specific patient i.e. that the CSCs for the preferential expression plots are taken from. So, the plot needs to fulfil the criteria that the CSCs are from only one patient, but the tumour cell population is from multiple patients.

HUGO_groups_sets_of_columnNumbers
List of lists, i.e. a list of HUGO gene groups, with each named gene group containing positions (row numbers) in the expression matrix tExprMatr of all the gene group members present in the expression matrix.

pCSC_group_name
Name of CSC group to print preferential expression plots for e.g. 'pCSC_group_1_SC4_Talazoparib.LB17011', but can be anything.

outDir
Output directory for plots.

tissue_type
Name of tissue type e.g. 'CDX'; can be anything.

family_type_name
Name of the gene group type; usually 'HUGO'.

patient_source_name
Optional: Name of the patient sample that CSC cells are from; just name for in the plot title, can be anything.

frag_count_method
Expression measurement type, e.g. 'raw'; but just a label, can be anything.

separate_row_for_high_expression_transcripts
Optional: Set to 1 to print very high expression transcripts or gene groups on a separate row in some of the bar charts; helpful when there is a big difference between the most highly expressed and those on the same bar chart row.

manually_set_max_number_of_graph_rows
Optional: Set to manually override the maximum number of rows shown in the stacked family plot.

Value

List of highest ranked preferentially expressed genes and gene families: - ranked families from family barcharts; - ranked transcripts from family barcharts.